# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "properties": {
          "decision": {
            "enum": [
              "approved",
              "denied"
            ],
            "title": "Decision",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoryItemOut": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entity_names": {
            "items": {
              "type": "string"
            },
            "title": "Entity Names",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "importance": {
            "title": "Importance",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "message_ids": {
            "items": {
              "type": "string"
            },
            "title": "Message Ids",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "kind",
          "content",
          "entity_names",
          "message_ids",
          "importance",
          "created_at",
          "updated_at"
        ],
        "title": "MemoryItemOut",
        "type": "object"
      },
...
      "ToolCallOut": {
        "properties": {
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "request_url": {
            "title": "Request Url",
            "type": "string"
          },
          "response_body": {
            "title": "Response Body",
            "type": "string"
          },
          "response_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Status"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tool_id": {
            "title": "Tool Id",
            "type": "string"
          },
          "tool_name": {
            "title": "Tool Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "conversation_id",
          "tool_id",
          "tool_name",
          "status",
          "request_url",
          "response_status",
          "response_body",
          "error",
          "created_at"
        ],
        "title": "ToolCallOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Durable chat, cited local knowledge, provenance, and permissioned read-only tools.",
    "title": "Agentic Knowledge Workspace API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/memory": {
      "get": {
        "operationId": "list_memory_api_memory_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemoryItemOut"
                  },
                  "title": "Response List Memory Api Memory Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Memory",
        "tags": [
          "memory"
        ]
      }
    },
    "/api/memory/{memory_id}": {
      "delete": {
        "operationId": "forget_memory_api_memory__memory_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forget Memory",
        "tags": [
          "memory"
        ]
      }
    },
...
    "/api/tool-calls": {
      "get": {
        "operationId": "list_tool_calls_api_tool_calls_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ToolCallOut"
                  },
                  "title": "Response List Tool Calls Api Tool Calls Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tool Calls",
        "tags": [
          "tools"
        ]
      }
    },
    "/api/tool-calls/{tool_call_id}/decision": {
      "post": {
        "operationId": "decide_tool_call_api_tool_calls__tool_call_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "tool_call_id",
            "required": true,
            "schema": {
              "title": "Tool Call Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-user-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000002",
              "title": "X-User-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "default": "00000000-0000-4000-8000-000000000001",
              "title": "X-Workspace-Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolCallOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Tool Call",
        "tags": [
          "tools"
        ]
      }
    },
...
  }
}